+++ mirrord/config/src/agent.rs
@@ -67,6 +67,13 @@ pub struct AgentConfig {
     /// and if that fails it uses eth0.
     #[config(env = "MIRRORD_AGENT_NETWORK_INTERFACE")]
     pub network_interface: Option<String>,
+
+    /// Controls target pause feature. Unstable.
+    ///
+    /// With this feature enabled, the remote container is paused while clients are connected to
+    /// the agent.
+    #[config(env = "MIRRORD_PAUSE", default = false, unstable)]
+    pub pause: bool,
 }
 
 #[cfg(test)]