+++ CHANGELOG.md
@@ -7,6 +7,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Added
+
+- `--pause` feature (unstable). See [#712](https://github.com/metalbear-co/mirrord/issues/712).
+- operator setup cli feature.
+
 ### Changed
 
 - CI: cancel previous runs of same PR.